update instance debug output with basic image information
you can now specify --debug-img as a cli argument and mhubio will automatically print image shape, size, origin as well as the min and max values. Note, that the performance is decreased by the additional read operations and no caching is performed as this is only an optional debug feature.