Migrate from setuptools to uv pyproject (#4)
* replace setup.py with pyproject.toml

* set include_package_data = true

* fix include-package-data

* specify packages.find.include

- trying to fix the issue where mhubio.core is not found when importing the build project

* specify package-dir for a flat-layout with additional top level folders

NOTE: we have a flat layout (all modules and sub modules are arranged in the mhubio/ folder) but also have additional top level folders (hence we need to specify what exactly to include). A manual setup is to be avoided since we have a nested structure that may grow further over time with additional modules and we don't want to manually specify all those subfolders. An alternative would  be to move from a (semi) flat layout to a src layout. For now, however, we aim to avoid any major restructorization.

* set package name (mhubio)

* bump version 0.0.2 -> 0.0.3